refactor(admin): split monolithic handlers.rs into sub-domain files
