+++ server/src/admin/mod.rs
@@ -4,12 +4,16 @@
 //! - Non-elevated: list users, list guilds, audit log, elevate/de-elevate session
 //! - Elevated: ban users, suspend guilds, manage announcements
 
+pub mod audit;
 pub mod error;
-pub mod handlers;
+pub mod guilds;
 pub mod middleware;
 pub mod observability;
 pub mod queries;
+pub(crate) mod shared;
+pub mod system;
 pub mod types;
+pub mod users;
 
 use axum::middleware::from_fn_with_state;
 use axum::routing::{delete, get, post, put};
@@ -107,53 +111,53 @@ pub fn router(state: AppState) -> Router<AppState> {
         // User management
         .route(
             "/users/{id}/ban",
-            post(handlers::ban_user).delete(handlers::unban_user),
+            post(users::ban_user).delete(users::unban_user),
         )
-        .route("/users/{id}/unban", post(handlers::unban_user))
-        .route("/users/bulk-ban", post(handlers::bulk_ban_users))
-        .route("/users/{id}", delete(handlers::delete_user))
+        .route("/users/{id}/unban", post(users::unban_user))
+        .route("/users/bulk-ban", post(users::bulk_ban_users))
+        .route("/users/{id}", delete(users::delete_user))
         .route(
             "/guilds/{id}/suspend",
-            post(handlers::suspend_guild).delete(handlers::unsuspend_guild),
+            post(guilds::suspend_guild).delete(guilds::unsuspend_guild),
         )
-        .route("/guilds/{id}/unsuspend", post(handlers::unsuspend_guild))
-        .route("/guilds/bulk-suspend", post(handlers::bulk_suspend_guilds))
-        .route("/guilds/{id}", delete(handlers::delete_guild))
-        .route("/announcements", post(handlers::create_announcement))
+        .route("/guilds/{id}/unsuspend", post(guilds::unsuspend_guild))
+        .route("/guilds/bulk-suspend", post(guilds::bulk_suspend_guilds))
+        .route("/guilds/{id}", delete(guilds::delete_guild))
+        .route("/announcements", post(system::create_announcement))
         // Auth settings (OIDC provider management)
         .route(
             "/auth-settings",
-            get(handlers::get_auth_settings).put(handlers::update_auth_settings),
+            get(system::get_auth_settings).put(system::update_auth_settings),
         )
         .route(
             "/oidc-providers",
-            get(handlers::list_oidc_providers).post(handlers::create_oidc_provider),
+            get(system::list_oidc_providers).post(system::create_oidc_provider),
         )
         .route(
             "/oidc-providers/{id}",
-            put(handlers::update_oidc_provider).delete(handlers::delete_oidc_provider),
+            put(system::update_oidc_provider).delete(system::delete_oidc_provider),
         )
         // Per-guild page limits
         .route(
             "/guilds/{id}/page-limits",
-            get(handlers::get_guild_page_limits).patch(handlers::set_guild_page_limits),
+            get(guilds::get_guild_page_limits).patch(guilds::set_guild_page_limits),
         )
         .layer(from_fn_with_state(state.clone(), require_elevated));
 
     // Non-elevated admin routes (require system admin)
     let admin_routes = Router::new()
         .route("/health", get(|| async { "admin ok" }))
-        .route("/stats", get(handlers::get_admin_stats))
-        .route("/users", get(handlers::list_users))
-        .route("/users/export", get(handlers::export_users_csv))
-        .route("/users/{id}/details", get(handlers::get_user_details))
-        .route("/guilds", get(handlers::list_guilds))
-        .route("/guilds/export", get(handlers::export_guilds_csv))
-        .route("/guilds/{id}/details", get(handlers::get_guild_details))
-        .route("/audit-log", get(handlers::get_audit_log))
+        .route("/stats", get(system::get_admin_stats))
+        .route("/users", get(users::list_users))
+        .route("/users/export", get(users::export_users_csv))
+        .route("/users/{id}/details", get(users::get_user_details))
+        .route("/guilds", get(guilds::list_guilds))
+        .route("/guilds/export", get(guilds::export_guilds_csv))
+        .route("/guilds/{id}/details", get(guilds::get_guild_details))
+        .route("/audit-log", get(audit::get_audit_log))
         .route(
             "/elevate",
-            post(handlers::elevate_session).delete(handlers::de_elevate_session),
+            post(system::elevate_session).delete(system::de_elevate_session),
         )
         .nest("/observability", observability::router())
         .merge(elevated_routes)
@@ -162,6 +166,6 @@ pub fn router(state: AppState) -> Router<AppState> {
     // Public admin routes (any authenticated user)
     // /status endpoint allows users to check their own admin status
     Router::new()
-        .route("/status", get(handlers::get_admin_status))
+        .route("/status", get(system::get_admin_status))
         .merge(admin_routes)
 }
