+++ server/src/openapi.rs
@@ -176,37 +176,37 @@ use utoipa::{Modify, OpenApi};
         crate::pages::handlers::get_guild_page,
         crate::pages::handlers::accept_guild_page,
         // Admin
-        crate::admin::handlers::get_admin_status,
-        crate::admin::handlers::get_admin_stats,
-        crate::admin::handlers::elevate_session,
-        crate::admin::handlers::de_elevate_session,
-        crate::admin::handlers::list_users,
-        crate::admin::handlers::export_users_csv,
-        crate::admin::handlers::get_user_details,
-        crate::admin::handlers::list_guilds,
-        crate::admin::handlers::export_guilds_csv,
-        crate::admin::handlers::get_guild_details,
-        crate::admin::handlers::get_audit_log,
+        crate::admin::system::get_admin_status,
+        crate::admin::system::get_admin_stats,
+        crate::admin::system::elevate_session,
+        crate::admin::system::de_elevate_session,
+        crate::admin::users::list_users,
+        crate::admin::users::export_users_csv,
+        crate::admin::users::get_user_details,
+        crate::admin::guilds::list_guilds,
+        crate::admin::guilds::export_guilds_csv,
+        crate::admin::guilds::get_guild_details,
+        crate::admin::audit::get_audit_log,
         crate::moderation::admin_handlers::list_reports,
         crate::moderation::admin_handlers::report_stats,
         crate::moderation::admin_handlers::get_report,
         crate::moderation::admin_handlers::claim_report,
         crate::moderation::admin_handlers::resolve_report,
-        crate::admin::handlers::ban_user,
-        crate::admin::handlers::unban_user,
-        crate::admin::handlers::bulk_ban_users,
-        crate::admin::handlers::delete_user,
-        crate::admin::handlers::suspend_guild,
-        crate::admin::handlers::unsuspend_guild,
-        crate::admin::handlers::bulk_suspend_guilds,
-        crate::admin::handlers::delete_guild,
-        crate::admin::handlers::create_announcement,
-        crate::admin::handlers::get_auth_settings,
-        crate::admin::handlers::update_auth_settings,
-        crate::admin::handlers::list_oidc_providers,
-        crate::admin::handlers::create_oidc_provider,
-        crate::admin::handlers::update_oidc_provider,
-        crate::admin::handlers::delete_oidc_provider,
+        crate::admin::users::ban_user,
+        crate::admin::users::unban_user,
+        crate::admin::users::bulk_ban_users,
+        crate::admin::users::delete_user,
+        crate::admin::guilds::suspend_guild,
+        crate::admin::guilds::unsuspend_guild,
+        crate::admin::guilds::bulk_suspend_guilds,
+        crate::admin::guilds::delete_guild,
+        crate::admin::system::create_announcement,
+        crate::admin::system::get_auth_settings,
+        crate::admin::system::update_auth_settings,
+        crate::admin::system::list_oidc_providers,
+        crate::admin::system::create_oidc_provider,
+        crate::admin::system::update_oidc_provider,
+        crate::admin::system::delete_oidc_provider,
         // Moderation
         crate::moderation::handlers::create_report,
         crate::moderation::filter_handlers::list_filter_configs,
@@ -459,24 +459,24 @@ use utoipa::{Modify, OpenApi};
         crate::admin::types::BulkSuspendRequest,
         crate::admin::types::BulkSuspendResponse,
         crate::admin::types::BulkActionFailure,
-        crate::admin::handlers::PaginationParams,
-        crate::admin::handlers::AuditLogParams,
-        crate::admin::handlers::UserSummary,
-        crate::admin::handlers::GuildSummary,
-        crate::admin::handlers::AuditLogEntryResponse,
-        crate::admin::handlers::DeElevateResponse,
-        crate::admin::handlers::UserGuildMembership,
-        crate::admin::handlers::UserDetailsResponse,
-        crate::admin::handlers::GuildMemberInfo,
-        crate::admin::handlers::GuildOwnerInfo,
-        crate::admin::handlers::GuildDetailsResponse,
-        crate::admin::handlers::PaginatedResponse<crate::admin::handlers::UserSummary>,
-        crate::admin::handlers::PaginatedResponse<crate::admin::handlers::GuildSummary>,
-        crate::admin::handlers::PaginatedResponse<crate::admin::handlers::AuditLogEntryResponse>,
-        crate::admin::handlers::DeleteResponse,
-        crate::admin::handlers::AnnouncementResponse,
-        crate::admin::handlers::AuthSettingsResponse,
-        crate::admin::handlers::OidcProviderResponse,
+        crate::admin::shared::PaginationParams,
+        crate::admin::audit::AuditLogParams,
+        crate::admin::users::UserSummary,
+        crate::admin::guilds::GuildSummary,
+        crate::admin::audit::AuditLogEntryResponse,
+        crate::admin::system::DeElevateResponse,
+        crate::admin::users::UserGuildMembership,
+        crate::admin::users::UserDetailsResponse,
+        crate::admin::guilds::GuildMemberInfo,
+        crate::admin::guilds::GuildOwnerInfo,
+        crate::admin::guilds::GuildDetailsResponse,
+        crate::admin::shared::PaginatedResponse<crate::admin::users::UserSummary>,
+        crate::admin::shared::PaginatedResponse<crate::admin::guilds::GuildSummary>,
+        crate::admin::shared::PaginatedResponse<crate::admin::audit::AuditLogEntryResponse>,
+        crate::admin::shared::DeleteResponse,
+        crate::admin::system::AnnouncementResponse,
+        crate::admin::system::AuthSettingsResponse,
+        crate::admin::system::OidcProviderResponse,
         // Social
         crate::social::types::FriendshipStatus,
         crate::social::types::Friendship,
